requirements.txt
pytest
selenium
jsonpickle
pony
pymysql
allure-pytest
